pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "kb-generator"
version = "0.1.0"
description = "Generate LLM-friendly knowledge bases from .NET/C# projects"
readme = "README.md"
requires-python = ">=3.11"
license = { text = "MIT" }

dependencies = [
    "click>=8.1",
    "pyyaml>=6.0",
    "jinja2>=3.1",
    "tree-sitter>=0.23.0,<0.25.0",
    "tree-sitter-c-sharp>=0.23.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-cov>=4.0",
]

[project.scripts]
kb-gen = "kb_generator.cli:cli"

[tool.setuptools.packages.find]
include = ["kb_generator*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
